refactor(user): 重构用户实体类名为Users并更新相关引用
- 将User实体类重命名为Users并更新数据库表名为users
- 修改Gtype实体中gtypeId字段类型从Short改为Integer
- 更新UserController中所有User类型引用为Users
- 更新UserRepository接口中的泛型类型为Users
- 更新UserService中所有User类型引用为Users
- 修正README.md中用户接口路径从/user改为/users

+++ src/main/java/com/trallkong/csitsbackend/entity/Gtype.java
@@ -20,7 +20,7 @@ public class Gtype extends BaseEntity {
     @Column(name = "gtype_id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonProperty("gtype_id")
-    private Short gtypeId;
+    private Integer gtypeId;
 
     // 商品类型名
     @Column(name = "type_name", nullable = false)
